requirements.txt
MetaTrader5>=5.0.45
python-dotenv>=1.0.0
loguru>=0.7.0
